fix: 게시글 및 일정 삭제 서비스에 예외처리 추가  (#148)
* fix: 일정 삭제 서비스에 예외처리 추가

* fix: 게시글 삭제 서비스에 예외처리 추가

+++ src/main/java/com/sammaru5/sammaru/service/schedule/ScheduleRemoveService.java
@@ -1,5 +1,7 @@
 package com.sammaru5.sammaru.service.schedule;
 
+import com.sammaru5.sammaru.exception.CustomException;
+import com.sammaru5.sammaru.exception.ErrorCode;
 import com.sammaru5.sammaru.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,10 @@ public class ScheduleRemoveService {
     private final ScheduleRepository scheduleRepository;
 
     public boolean removeScheduleById(Long scheduleId) {
+        if (!scheduleRepository.existsById(scheduleId)) {
+            throw new CustomException(ErrorCode.SCHEDULE_NOT_FOUND, scheduleId.toString());
+        }
+
         scheduleRepository.deleteById(scheduleId);
         return true;
     }
